requirements.txt
chardet==5.2.0
contourpy==1.2.0
cycler==0.12.1
filelock==3.13.3
fonttools==4.50.0
fsspec==2024.3.1
furl==2.1.3
greenlet==3.0.3
Jinja2==3.1.3
kiwisolver==1.4.5
logging-tree==1.9
logx==0.1.1579232358
Markdown==3.6
MarkupSafe==2.1.5
matplotlib==3.8.3
migra==3.0.1663481299
mpmath==1.3.0
networkx==3.2.1
numpy==1.26.4
orderedmultidict==1.0.1
packaging==24.0
pendulum==3.0.0
pillow==10.2.0
psycopg2-binary==2.9.9
pyparsing==3.1.2
python-dateutil==2.9.0.post0
PyYAML==6.0.1
results==0.1.1588569394
schemainspect==3.1.1663587362
six==1.16.0
SQLAlchemy==2.0.29
sqlbag==0.1.1617247075
sympy==1.12
tabulate==0.9.0
time-machine==2.14.1
torch==2.2.1
torchvision==0.17.1
typing_extensions==4.10.0
tzdata==2024.1
